Restructure to files to match prod
