+++ spark/components/card/react/SprkCard.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import SprkCardHighlightedHeader from './SprkCardHighlightedHeader';
-import SprkCardTeaser from './SprkCardTeaser';
+import SprkCardHighlightedHeader from './components/SprkCardHighlightedHeader/SprkCardHighlightedHeader';
+import SprkCardTeaser from './components/SprkCardTeaser/SprkCardTeaser';
 
 const SprkCard = (props) => {
   const {
