+++ spark/components/card/react/components/SprkCardHighlightedHeader/SprkCardHighlightedHeader.js
