+++ spark/components/card/react/components/SprkCardTeaser/SprkCardTeaser.test.js
